+++ demo/guihello/seltext.c
@@ -11,7 +11,6 @@ typedef struct _sel_data_t SelData;
 struct _sel_data_t
 {
     Window *window;
-    Edit *edit;
     Edit *edit_range;
     Button *caps;
     TextView *text;
@@ -33,10 +32,10 @@ static void i_OnCopy(SelData *data, Event *e)
     cassert_no_null(data);
     unref(e);
     control = window_get_focus(data->window);
-    if (guicontrol_edit(control) == data->edit)
-        edit_copy(data->edit);
-    else if (guicontrol_textview(control) == data->text)
-        textview_copy(data->text);
+    if (guicontrol_edit(control) != NULL)
+        edit_copy(guicontrol_edit(control));
+    else if (guicontrol_textview(control) != NULL)
+        textview_copy(guicontrol_textview(control));
 }
 
 /*---------------------------------------------------------------------------*/
@@ -47,10 +46,10 @@ static void i_OnPaste(SelData *data, Event *e)
     cassert_no_null(data);
     unref(e);
     control = window_get_focus(data->window);
-    if (guicontrol_edit(control) == data->edit)
-        edit_paste(data->edit);
-    else if (guicontrol_textview(control) == data->text)
-        textview_paste(data->text);
+    if (guicontrol_edit(control) != NULL)
+        edit_paste(guicontrol_edit(control));
+    else if (guicontrol_textview(control) != NULL)
+        textview_paste(guicontrol_textview(control));
 }
 
 /*---------------------------------------------------------------------------*/
@@ -61,10 +60,10 @@ static void i_OnCut(SelData *data, Event *e)
     cassert_no_null(data);
     unref(e);
     control = window_get_focus(data->window);
-    if (guicontrol_edit(control) == data->edit)
-        edit_cut(data->edit);
-    else if (guicontrol_textview(control) == data->text)
-        textview_cut(data->text);
+    if (guicontrol_edit(control) != NULL)
+        edit_cut(guicontrol_edit(control));
+    else if (guicontrol_textview(control) != NULL)
+        textview_cut(guicontrol_textview(control));
 }
 
 /*---------------------------------------------------------------------------*/
@@ -83,10 +82,10 @@ static void i_OnSelect(SelData *data, Event *e)
     start = str_to_i32(tc(left), 10, NULL);
     end = str_to_i32(tc(right), 10, NULL);
 
-    if (guicontrol_edit(control) == data->edit)
-        edit_select(data->edit, start, end);
-    else if (guicontrol_textview(control) == data->text)
-        textview_select(data->text, start, end);
+    if (guicontrol_edit(control) != NULL)
+        edit_select(guicontrol_edit(control), start, end);
+    else if (guicontrol_textview(control) != NULL)
+        textview_select(guicontrol_textview(control), start, end);
 
     str_destroy(&left);
     str_destroy(&right);
@@ -103,7 +102,7 @@ static void i_OnWrap(SelData *data, Event *e)
 
 /*---------------------------------------------------------------------------*/
 
-static void i_OnEditFilter(SelData *data, Event *e)
+static void i_filter_event(SelData *data, Event *e, const char_t *name)
 {
     const EvText *p = event_params(e, EvText);
     cassert_no_null(data);
@@ -144,7 +143,24 @@ static void i_OnEditFilter(SelData *data, Event *e)
         r->apply = TRUE;
     }
 
-    textview_printf(data->info_text, "Edit: Pos %d Len %d\n", p->cpos, p->len);
+    textview_printf(data->info_text, "%s: Pos %d Len %d\n", name, p->cpos, p->len);
+    textview_scroll_caret(data->info_text);
+}
+
+/*---------------------------------------------------------------------------*/
+
+static void i_OnEditFilter(SelData *data, Event *e)
+{
+    i_filter_event(data, e, "Edit OnFilter");
+}
+
+/*---------------------------------------------------------------------------*/
+
+static void i_OnEditChange(SelData *data, Event *e)
+{
+    const EvText *p = event_params(e, EvText);
+    cassert_no_null(data);
+    textview_printf(data->info_text, "Edit OnChange: Pos: %d Len %d\n", p->cpos, p->len);
     textview_scroll_caret(data->info_text);
 }
 
@@ -152,46 +168,7 @@ static void i_OnEditFilter(SelData *data, Event *e)
 
 static void i_OnTextViewFilter(SelData *data, Event *e)
 {
-    const EvText *p = event_params(e, EvText);
-    cassert_no_null(data);
-
-    /*
-     * Convert the inserted text in caps.
-     * p->text the inserted text only (const).
-     * r->text the new filterd text.
-     * p->cpos current caret position.
-     * r->apply = TRUE means the editbox text has to be updated.
-     * p->len number of chars inserted at left of caret (to caps).
-     * We are working with UTF8-Strings. Sequential access.
-     */
-    if (button_get_state(data->caps) == ekGUI_ON && p->len > 0)
-    {
-        EvTextFilter *r = event_result(e, EvTextFilter);
-        const char_t *src = p->text;
-        char_t *dest = r->text;
-        uint32_t cp = unicode_to_u32(src, ekUTF8);
-        uint32_t pos = 0;
-        while (cp != 0)
-        {
-            uint32_t offset = 0;
-
-            if (cp >= 'a' && cp <= 'z')
-                cp -= 32;
-
-            offset = unicode_to_char(cp, dest, ekUTF8);
-            dest += offset;
-            src = unicode_next(src, ekUTF8);
-            cp = unicode_to_u32(src, ekUTF8);
-            pos += 1;
-        }
-
-        *dest = 0;
-        r->cpos = p->cpos;
-        r->apply = TRUE;
-    }
-
-    textview_printf(data->info_text, "TextView: Pos %d Len %d\n", p->cpos, p->len);
-    textview_scroll_caret(data->info_text);
+    i_filter_event(data, e, "TextView OnFilter");
 }
 
 /*---------------------------------------------------------------------------*/
@@ -245,37 +222,55 @@ static Layout *i_text_controls(SelData *data)
 
 static Layout *i_layout(SelData *data)
 {
-    Layout *layout1 = layout_create(1, 7);
+    Layout *layout1 = layout_create(1, 11);
     Layout *layout2 = i_text_controls(data);
     Label *label1 = label_create();
     Label *label2 = label_create();
     Label *label3 = label_create();
-    Edit *edit = edit_create();
+    Label *label4 = label_create();
+    Label *label5 = label_create();
+    Edit *edit1 = edit_create();
+    Edit *edit2 = edit_multiline();
+    Edit *edit3 = edit_multiline();
     TextView *text1 = textview_create();
     TextView *text2 = textview_create();
     cassert_no_null(data);
     label_text(label1, "EditBox");
-    label_text(label2, "TextView");
-    label_text(label3, "Info");
-    edit_text(edit, "This is a text in the EditBox control");
-    edit_OnFilter(edit, listener(data, i_OnEditFilter, SelData));
+    label_text(label2, "EditBox (multiline)");
+    label_text(label3, "EditBox (multiline with user-height)");
+    label_text(label4, "TextView");
+    label_text(label5, "Info");
+    edit_text(edit1, "This is a text in the EditBox control");
+    edit_text(edit2, "This is a text in the multiline EditBox control");
+    edit_text(edit3, "This is a text in the multiline EditBox control");
+    edit_min_height(edit3, 100);
+    edit_OnFilter(edit1, listener(data, i_OnEditFilter, SelData));
+    edit_OnFilter(edit2, listener(data, i_OnEditFilter, SelData));
+    edit_OnFilter(edit3, listener(data, i_OnEditFilter, SelData));
+    edit_OnChange(edit1, listener(data, i_OnEditChange, SelData));
+    edit_OnChange(edit2, listener(data, i_OnEditChange, SelData));
+    edit_OnChange(edit3, listener(data, i_OnEditChange, SelData));
     textview_writef(text1, "This is another text in the TextView control, wider than the control.");
     textview_editable(text1, TRUE);
     textview_OnFilter(text1, listener(data, i_OnTextViewFilter, SelData));
     layout_label(layout1, label1, 0, 0);
-    layout_edit(layout1, edit, 0, 1);
+    layout_edit(layout1, edit1, 0, 1);
     layout_label(layout1, label2, 0, 2);
-    layout_textview(layout1, text1, 0, 3);
-    layout_layout(layout1, layout2, 0, 4);
-    layout_label(layout1, label3, 0, 5);
-    layout_textview(layout1, text2, 0, 6);
-    layout_tabstop(layout1, 0, 3, TRUE);
-    layout_tabstop(layout1, 0, 4, FALSE);
-    layout_halign(layout1, 0, 4, ekLEFT);
+    layout_edit(layout1, edit2, 0, 3);
+    layout_label(layout1, label3, 0, 4);
+    layout_edit(layout1, edit3, 0, 5);
+    layout_label(layout1, label4, 0, 6);
+    layout_textview(layout1, text1, 0, 7);
+    layout_layout(layout1, layout2, 0, 8);
+    layout_label(layout1, label5, 0, 9);
+    layout_textview(layout1, text2, 0, 10);
+    layout_tabstop(layout1, 0, 7, TRUE);
+    layout_tabstop(layout1, 0, 8, FALSE);
+    layout_halign(layout1, 0, 8, ekLEFT);
     layout_vmargin(layout1, 1, 5);
     layout_vmargin(layout1, 3, 5);
-    layout_vmargin(layout1, 4, 5);
-    data->edit = edit;
+    layout_vmargin(layout1, 5, 5);
+    layout_vmargin(layout1, 7, 5);
     data->text = text1;
     data->info_text = text2;
     return layout1;